Add files via upload
Here we begin a rough outline of the general data structure and application of Propagating Structural Bifurcations along with Splits within the SERD network. We show that for two PSBs propagating in two opposite directions that the resultants are isomorphic to one another under changes in the order of PSB propagation through computational time.