Create SpanningTreeFromSplittingA1
This code implements simple duplication expansion and splitting of the radial subhypergraph of the SERD network. Reductions are not included. It outputs a hypergraph edge set evoltuion and generates a spanning tree, which is similar to mycelial strucutre resulting from inward propagating strucutral bifurcations against the outward pressure of duplications.