GrowingSERDnetworkWithInfoProp
In this simulation two evolution sets are output, one for the hypergraphevolution and another for the location of the propagating information. These can be combined to create an evolution of the radial network with inward information propagation. Each vertex in the PIPev set can be coloured to represent that it stores a bit of information. Information dilution and redshift may be observed. Reductions are not included and so this will only give the averaged behaviour of energy dilution due to spatial expansion and will not incorperate all aspects of the SERD network.