Create TestOfRedShiftScaleFactorRelationInSERDModelPythonCode
This simulation scales a single SE filament with an initial Monte Carlo (MC) loop, allowing for a sample of scaled inputs to be put into a second MC loop. When PIPs update on an observing PP, in this second scaled MC loop, data of time of emission and observation as well as the corresponding scale factors are stored along side the gamma value, corresponding to its information density or number of unit PIPs that make it up. For the model to match the simulation then the average output gamma from the mathematical model must approach the value of the gamma in the simulation. It should approach the line y=x, as the system is scaled up, for the model to corroborate the simulation. This is what is observed.