Add files via upload
This algorithm demonstrates a correlation between information propagation and interaction from one end of a growing filament of Space Elements and the observed cosmological red-shift scale factor relation. Single bits of information, represent by a value of 1, flows into the filament from one end ever TS and propagates along the filament at a rate of 1 Space Element per TS as the filament stochastically evolves. This causes dilution and superposition of filaments as they propagates leading to Propagating Information Packets (PIP). If the number of bits that makes up a PIP at observation is related to the energy of the PIP, and a PIP is analogous to a photon, the red-shifting effect due to system scaling is observed and matches that of the cosmological redshift scale factor relation, connecting an unobserved information propagation mechanism with the energy of a photon.

This algorithm scales an set of evolutions to an input set that is scaled via Monte-Carlo simulation. These new scaled inputs are randomly selected as initial conditions for a second MC loop. This is then run and the gamma values (PIP magnitudes) at observation and emission are stored. The algorithm tests the model against the cosmological redshift scale factor relation.